+++ src/main/java/client/command/commands/gm4/BossDropRateCommand.java
@@ -40,7 +40,7 @@ public class BossDropRateCommand extends Command {
             return;
         }
 
-        int bossdroprate = Math.max(Integer.parseInt(params[0]), 1);
+        float bossdroprate = Math.max(Float.parseFloat(params[0]), 1.0f);
         c.getWorldServer().setBossDropRate(bossdroprate);
         c.getWorldServer().broadcastPacket(PacketCreator.serverNotice(6, "[Rate] Boss Drop Rate has been changed to " + bossdroprate + "x."));
     }
