+++ src/main/java/config/WorldConfig.java
@@ -6,11 +6,11 @@ public class WorldConfig {
     public String event_message = "";
     public String why_am_i_recommended = "";
     public int channels = 1;
-    public int exp_rate = 1;
-    public int meso_rate = 1;
-    public int drop_rate = 1;
-    public int boss_drop_rate = 1;
-    public int quest_rate = 1;
-    public int travel_rate = 1;
-    public int fishing_rate = 1;
+    public float exp_rate = 1.0f;
+    public float meso_rate = 1.0f;
+    public float drop_rate = 1.0f;
+    public float boss_drop_rate = 1.0f;
+    public float quest_rate = 1.0f;
+    public float travel_rate = 1.0f;
+    public float fishing_rate = 1.0f;
 }
