+++ src/main/java/scripting/event/EventInstanceManager.java
@@ -188,12 +188,12 @@ public class EventInstanceManager {
 
         if (mapId == -1) {
             for (Character mc : players) {
-                mc.gainExp(gain * mc.getExpRate(), true, true);
+                mc.gainExp((int) (gain * mc.getExpRate()), true, true);
             }
         } else {
             for (Character mc : players) {
                 if (mc.getMapId() == mapId) {
-                    mc.gainExp(gain * mc.getExpRate(), true, true);
+                    mc.gainExp((int) (gain * mc.getExpRate()), true, true);
                 }
             }
         }
@@ -212,12 +212,12 @@ public class EventInstanceManager {
 
         if (mapId == -1) {
             for (Character mc : players) {
-                mc.gainMeso(gain * mc.getMesoRate());
+                mc.gainMeso((int) (gain * mc.getMesoRate()));
             }
         } else {
             for (Character mc : players) {
                 if (mc.getMapId() == mapId) {
-                    mc.gainMeso(gain * mc.getMesoRate());
+                    mc.gainMeso((int) (gain * mc.getMesoRate()));
                 }
             }
         }
