+++ src/main/java/server/quest/actions/ExpAction.java
@@ -52,9 +52,9 @@ public class ExpAction extends AbstractQuestAction {
 
     public static void runAction(Character chr, int gain) {
         if (!YamlConfig.config.server.USE_QUEST_RATE) {
-            chr.gainExp(gain * chr.getExpRate(), true, true);
+            chr.gainExp((int) (gain * chr.getExpRate()), true, true);
         } else {
-            chr.gainExp(gain * chr.getQuestExpRate(), true, true);
+            chr.gainExp((int) (gain * chr.getQuestExpRate()), true, true);
         }
     }
 }
